xnualloc: translate ordinary zone paths and wire opt-in Vinix heap
Add externally serialized non-SMR zone allocation/import/drop, full/partial/
empty queues, bitmap and live-object accounting, cache refill/overflow,
depot recirculation and draining. Cached objects remain reserved until drain.

Wire malloc/free/realloc/heap_trim behind -d xnu_zone. Publish CPU-cache
readiness only after SMP installs CPU numbers. Bootstrap and CPU IDs >=64
use uncached allocations. Cache metadata uses PMM rather than recursive
malloc and is retained; allocation failure falls back to uncached operation.
Require available payload slots before allocating new cache metadata.

Keep one truly empty spare per class; detach reclaimable pages before
unlock/free. Preserve zeroing, poisoning and existing large-allocation ABI.
All CPU caches are protected by the same class lock: this is NOT XNU's
scalable CPU fast path, SMR implementation, typed heaps or Mach VM port.

Add five actual V state-machine tests (not executed locally). Local
Python protocol model passed six tests including 80,000 mixed operations;
prior model/reference suites passed 10 and 4 tests respectively. No local
V compiler or QEMU is available; V/kernel CI was queued at the last check.
Retain draft status; no performance or production-readiness claim.

+++ kernel/modules/xnualloc/magazine.v
@@ -63,11 +63,11 @@
 @[manualfree]
 module xnualloc
 
-// Data-structure translations only. These are NOT enabled as Vinix per-CPU
-// caches. Their callers must supply CPU pinning, locks, ownership transitions,
-// refill, accounting and safe reclamation. No SMR grace-period implementation
-// is hidden behind a successful stub. Depot polling requires an explicit poll
-// callback when SMR is requested.
+// Data-structure translations used by zone.v and its opt-in Vinix adapter.
+// Callers supply CPU pinning, locks, ownership transitions, refill, accounting
+// and safe reclamation. No SMR grace-period implementation is hidden behind
+// a successful stub. Depot polling requires an explicit poll callback when
+// SMR is requested; the zone.v backend is strictly non-SMR.
 //
 // Storage adaptation: the flexible C array has a fixed 32-element capacity.
 pub const magazine_capacity = u16(32)
@@ -178,7 +178,7 @@ pub fn (mut d Depot) zone_depot_pop_head_empty(minimum &RecircMinimum) &Magazine
 // Source and destination must be distinct, initialized, externally locked.
 pub fn (mut dst Depot) zone_depot_move_full(mut src Depot, n u32,
 	minimum &RecircMinimum, lifo bool) u64 {
-	assert n > 0 && src.full >= n && &dst != &src
+	assert n > 0 && src.full >= n && u64(&dst) != u64(&src)
 	src.full -= n
 	unsafe {
 		if minimum != nil && minimum.full > src.full {
@@ -212,7 +212,7 @@ pub fn (mut dst Depot) zone_depot_move_full(mut src Depot, n u32,
 
 pub fn (mut dst Depot) zone_depot_move_empty(mut src Depot, n u32,
 	minimum &RecircMinimum) {
-	assert n > 0 && src.empty >= n && &dst != &src
+	assert n > 0 && src.empty >= n && u64(&dst) != u64(&src)
 	src.empty -= n
 	unsafe {
 		if minimum != nil && minimum.empty > src.empty {
